backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
langchain>=0.0.300
langchain-openai>=0.0.1
langchain-anthropic>=0.0.1
langchain-community>=0.0.1
openai>=1.0.0
anthropic>=0.17.0
httpx>=0.24.0
python-multipart>=0.0.6
sse-starlette>=1.6.0
typing-extensions>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-docx>=1.1.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
markdown>=3.5.0 